+++ src/cross-client.ts
@@ -74,6 +74,11 @@ export class CrossClient {
   }
   private _useVersionedTxs: boolean = false;
 
+  public get txRetryAmount(): number {
+    return this._txRetryAmount;
+  }
+  private _txRetryAmount: number = undefined;
+
   /**
    * Client usdc account address.
    */
@@ -275,6 +280,7 @@ export class CrossClient {
     throttle: boolean = false,
     delegator: PublicKey = undefined,
     useVersionedTxs: boolean = false,
+    txRetryAmount: number = undefined,
     subaccountIndex: number = 0
   ): Promise<CrossClient> {
     let client = new CrossClient(connection, wallet, opts);
@@ -286,6 +292,7 @@ export class CrossClient {
 
     client._subaccountIndex = subaccountIndex;
     client._useVersionedTxs = useVersionedTxs;
+    client._txRetryAmount = txRetryAmount;
     client._usdcAccountAddress = utils.getAssociatedTokenAddress(
       Exchange.usdcMintAddress,
       owner
@@ -526,6 +533,10 @@ export class CrossClient {
     this._useVersionedTxs = useVersionedTxs;
   }
 
+  public setTxRetryAmount(txRetryAmount: number) {
+    this._txRetryAmount = txRetryAmount;
+  }
+
   /**
    * @param timerInterval   desired interval for client polling.
    */
@@ -663,7 +674,9 @@ export class CrossClient {
           undefined,
           undefined,
           undefined,
-          this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+          this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+          undefined,
+          this._txRetryAmount
         )
       );
     }
@@ -748,7 +761,9 @@ export class CrossClient {
           undefined,
           undefined,
           undefined,
-          this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+          this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+          undefined,
+          this._txRetryAmount
         )
       );
     }
@@ -829,7 +844,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
@@ -859,7 +876,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
     this._account = null;
     return txId;
@@ -886,8 +905,47 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
+    );
+    this._accountManager = null;
+    return txId;
+  }
+
+  /**
+   * Closes the CrossMarginAccount and CrossMarginAccountManager in one transaction
+   */
+  public async closeAccountAndManager(): Promise<TransactionSignature> {
+    this.delegatedCheck();
+    if (this._account === null) {
+      throw Error("User has no account to close");
+    }
+    if (this._accountManager === null) {
+      throw Error("User has no account manager to close");
+    }
+    let tx = new Transaction().add(
+      instructions.closeCrossMarginAccountIx(
+        this._provider.wallet.publicKey,
+        this._accountAddress,
+        this._accountManagerAddress
+      )
+    );
+    tx.add(
+      instructions.closeCrossMarginAccountManagerIx(
+        this._provider.wallet.publicKey,
+        this._accountManagerAddress
+      )
+    );
+    let txId: TransactionSignature = await utils.processTransaction(
+      this._provider,
+      tx,
+      undefined,
+      undefined,
+      undefined,
       this._useVersionedTxs ? utils.getZetaLutArr() : undefined
     );
+    this._accountManager = null;
     this._account = null;
     return txId;
   }
@@ -928,27 +986,48 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
   /**
    * Withdraws the entirety of the CrossClient's margin account and then closes it.
+   * Useful for only closing one subaccount.
    */
   public async withdrawAndCloseAccount(): Promise<TransactionSignature> {
     this.delegatedCheck();
     if (this._account === null) {
       throw Error("User has no margin account to withdraw or close.");
     }
+
     let tx = new Transaction();
-    tx.add(
-      instructions.withdrawV2Ix(
-        this._account.balance.toNumber(),
-        this._accountAddress,
-        this._usdcAccountAddress,
-        this._provider.wallet.publicKey
-      )
-    );
+    try {
+      await this.usdcAccountCheck();
+    } catch (e) {
+      tx.add(
+        splToken.Token.createAssociatedTokenAccountInstruction(
+          splToken.ASSOCIATED_TOKEN_PROGRAM_ID,
+          splToken.TOKEN_PROGRAM_ID,
+          constants.USDC_MINT_ADDRESS[Exchange.network],
+          this._usdcAccountAddress,
+          this.publicKey,
+          this.publicKey
+        )
+      );
+    }
+
+    if (this._account.balance > 0) {
+      tx.add(
+        instructions.withdrawV2Ix(
+          this._account.balance.toNumber(),
+          this._accountAddress,
+          this._usdcAccountAddress,
+          this._provider.wallet.publicKey
+        )
+      );
+    }
     tx.add(
       instructions.closeCrossMarginAccountIx(
         this._provider.wallet.publicKey,
@@ -962,9 +1041,76 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
+    );
+    this._account = null;
+    return txId;
+  }
+
+  /**
+   * Withdraws the entirety of the CrossClient's margin account and then closes it.
+   * Useful for closing the main account and everything
+   */
+  public async withdrawAndCloseAccountAndCloseManager(): Promise<TransactionSignature> {
+    this.delegatedCheck();
+    if (this._account === null) {
+      throw Error("User has no margin account to withdraw or close.");
+    }
+    if (this._accountManager === null) {
+      throw Error("User has no account manager to close");
+    }
+
+    let tx = new Transaction();
+    try {
+      await this.usdcAccountCheck();
+    } catch (e) {
+      tx.add(
+        splToken.Token.createAssociatedTokenAccountInstruction(
+          splToken.ASSOCIATED_TOKEN_PROGRAM_ID,
+          splToken.TOKEN_PROGRAM_ID,
+          constants.USDC_MINT_ADDRESS[Exchange.network],
+          this._usdcAccountAddress,
+          this.publicKey,
+          this.publicKey
+        )
+      );
+    }
+
+    if (this._account.balance > 0) {
+      tx.add(
+        instructions.withdrawV2Ix(
+          this._account.balance.toNumber(),
+          this._accountAddress,
+          this._usdcAccountAddress,
+          this._provider.wallet.publicKey
+        )
+      );
+    }
+    tx.add(
+      instructions.closeCrossMarginAccountIx(
+        this._provider.wallet.publicKey,
+        this._accountAddress,
+        this._accountManagerAddress
+      )
+    );
+    tx.add(
+      instructions.closeCrossMarginAccountManagerIx(
+        this._provider.wallet.publicKey,
+        this._accountManagerAddress
+      )
+    );
+    let txId: TransactionSignature = await utils.processTransaction(
+      this._provider,
+      tx,
+      undefined,
+      undefined,
+      undefined,
       this._useVersionedTxs ? utils.getZetaLutArr() : undefined
     );
     this._account = null;
+    this._accountManager = null;
     return txId;
   }
 
@@ -1081,7 +1227,8 @@ export class CrossClient {
       undefined,
       undefined,
       this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
-      options.blockhash
+      options.blockhash,
+      this._txRetryAmount
     );
     this._openOrdersAccounts[assetIndex] = openOrdersPda;
     return txId;
@@ -1219,7 +1366,8 @@ export class CrossClient {
       undefined,
       undefined,
       this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
-      options.blockhash
+      options.blockhash,
+      this._txRetryAmount
     );
     this._openOrdersAccounts[assetIndex] = openOrdersPda;
     return txId;
@@ -1276,6 +1424,59 @@ export class CrossClient {
     );
   }
 
+  /**
+   * Close all positions using market orders
+   * @param orderPrices Manual override for what price to send, decimal format
+   */
+  public async closeAllPositions(orderPrices: Map<Asset, number>) {
+    let ixs = [];
+
+    for (var position of [...this._positions.values()].flat()) {
+      let side = position.size < 0 ? types.Side.BID : types.Side.ASK;
+
+      if (!orderPrices.has(position.asset)) {
+        throw `orderPrices does not have a value for ${position.asset}`;
+      }
+      let price = orderPrices.get(position.asset);
+
+      ixs.push(
+        this.createPlacePerpOrderInstruction(
+          position.asset,
+          utils.convertDecimalToNativeInteger(price),
+          utils.convertDecimalToNativeLotSize(Math.abs(position.size)),
+          side,
+          { orderType: types.OrderType.FILLORKILL, tifOptions: {} } // We want to error if the trade didn't work for full size
+        )
+      );
+    }
+
+    let txs = utils.splitIxsIntoTx(
+      ixs,
+      this.useVersionedTxs
+        ? constants.MAX_ORDERS_PER_TX_LUT
+        : constants.MAX_ORDERS_PER_TX
+    );
+    let txIds: string[] = [];
+    await Promise.all(
+      txs.map(async (tx) => {
+        txIds.push(
+          await utils.processTransaction(
+            this._provider,
+            tx,
+            undefined,
+            undefined,
+            undefined,
+            this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+            undefined,
+            this._txRetryAmount
+          )
+        );
+      })
+    );
+
+    return txIds;
+  }
+
   public async cancelAllTriggerOrders(asset: Asset | undefined) {
     let orders = [];
     if (asset == undefined) {
@@ -1306,7 +1507,9 @@ export class CrossClient {
             undefined,
             undefined,
             undefined,
-            this.useVersionedTxs ? utils.getZetaLutArr() : undefined
+            this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+            undefined,
+            this._txRetryAmount
           )
         );
       })
@@ -1368,7 +1571,9 @@ export class CrossClient {
             undefined,
             undefined,
             undefined,
-            this.useVersionedTxs ? utils.getZetaLutArr() : undefined
+            this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+            undefined,
+            this._txRetryAmount
           )
         );
       })
@@ -1380,7 +1585,9 @@ export class CrossClient {
         undefined,
         undefined,
         undefined,
-        this.useVersionedTxs ? utils.getZetaLutArr() : undefined
+        this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+        undefined,
+        this._txRetryAmount
       )
     );
     return txIds;
@@ -1440,7 +1647,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
@@ -1493,7 +1702,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
@@ -1557,7 +1768,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
 
     // Editing a trigger order doesn't prompt an account change because the bits stay the same.
@@ -1591,7 +1804,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
@@ -1631,7 +1846,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
@@ -1666,7 +1883,18 @@ export class CrossClient {
     let txIds: string[] = [];
     await Promise.all(
       txs.map(async (tx) => {
-        txIds.push(await utils.processTransaction(this._provider, tx));
+        txIds.push(
+          await utils.processTransaction(
+            this._provider,
+            tx,
+            undefined,
+            undefined,
+            undefined,
+            this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+            undefined,
+            this._txRetryAmount
+          )
+        );
       })
     );
 
@@ -1710,7 +1938,18 @@ export class CrossClient {
     let txIds: string[] = [];
     await Promise.all(
       txs.map(async (tx) => {
-        txIds.push(await utils.processTransaction(this._provider, tx));
+        txIds.push(
+          await utils.processTransaction(
+            this._provider,
+            tx,
+            undefined,
+            undefined,
+            undefined,
+            this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+            undefined,
+            this._txRetryAmount
+          )
+        );
       })
     );
     return txIds;
@@ -1742,7 +1981,18 @@ export class CrossClient {
     let txIds: string[] = [];
     await Promise.all(
       txs.map(async (tx) => {
-        txIds.push(await utils.processTransaction(this._provider, tx));
+        txIds.push(
+          await utils.processTransaction(
+            this._provider,
+            tx,
+            undefined,
+            undefined,
+            undefined,
+            this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+            undefined,
+            this._txRetryAmount
+          )
+        );
       })
     );
     return txIds;
@@ -1817,7 +2067,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
@@ -1849,7 +2101,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
@@ -1917,7 +2171,8 @@ export class CrossClient {
       undefined,
       undefined,
       this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
-      options.blockhash
+      options.blockhash,
+      this._txRetryAmount
     );
   }
 
@@ -1984,7 +2239,8 @@ export class CrossClient {
       undefined,
       undefined,
       this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
-      newOptions.blockhash
+      newOptions.blockhash,
+      this._txRetryAmount
     );
   }
 
@@ -2052,7 +2308,8 @@ export class CrossClient {
       undefined,
       undefined,
       this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
-      newOptions.blockhash
+      newOptions.blockhash,
+      this._txRetryAmount
     );
   }
 
@@ -2082,7 +2339,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
     this._openOrdersAccounts[assetIndex] = openOrdersPda;
     return txId;
@@ -2128,7 +2387,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
     this._openOrdersAccounts[assetIndex] = PublicKey.default;
     return txId;
@@ -2143,18 +2404,13 @@ export class CrossClient {
   ): Promise<TransactionSignature[]> {
     this.delegatedCheck();
     let combinedIxs: TransactionInstruction[] = [];
-    let markets = [];
     for (var asset of assets) {
-      markets.push(Exchange.getPerpMarket(asset));
-    }
-    for (var i = 0; i < assets.length; i++) {
       let assetIndex = assetToIndex(asset);
-      let market = markets[i];
       if (this._openOrdersAccounts[assetIndex].equals(PublicKey.default)) {
         throw Error("User has no open orders account for this market!");
       }
       const [vaultOwner, _vaultSignerNonce] = utils.getSerumVaultOwnerAndNonce(
-        market,
+        Exchange.getPerpMarket(asset).address,
         constants.DEX_PID[Exchange.network]
       );
       let settleIx = instructions.settleDexFundsIx(
@@ -2163,7 +2419,7 @@ export class CrossClient {
         this._openOrdersAccounts[assetIndex]
       );
       let closeIx = instructions.closeOpenOrdersV3Ix(
-        market,
+        asset,
         this._provider.wallet.publicKey,
         this.accountAddress,
         this._openOrdersAccounts[assetIndex]
@@ -2187,7 +2443,9 @@ export class CrossClient {
         undefined,
         undefined,
         undefined,
-        this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+        this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+        undefined,
+        this._txRetryAmount
       );
       txIds.push(txId);
     }
@@ -2260,7 +2518,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
@@ -2320,7 +2580,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
@@ -2351,7 +2613,9 @@ export class CrossClient {
       undefined,
       undefined,
       undefined,
-      this._useVersionedTxs ? utils.getZetaLutArr() : undefined
+      this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+      undefined,
+      this._txRetryAmount
     );
   }
 
@@ -2434,7 +2698,9 @@ export class CrossClient {
             undefined,
             undefined,
             undefined,
-            this.useVersionedTxs ? utils.getZetaLutArr() : undefined
+            this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+            undefined,
+            this._txRetryAmount
           )
         );
       })
@@ -2479,7 +2745,9 @@ export class CrossClient {
             undefined,
             undefined,
             undefined,
-            this.useVersionedTxs ? utils.getZetaLutArr() : undefined
+            this._useVersionedTxs ? utils.getZetaLutArr() : undefined,
+            undefined,
+            this._txRetryAmount
           )
         );
       })